requirements.txt
python>=3.7
numpy
pillow
pyside2